# docs/openapi.yaml
openapi: "3.0.3"
info:
  title: RideShare API
  description: |
    Enterprise ride-sharing platform API.  
    All protected endpoints require a `Bearer` JWT in the `Authorization` header.
  version: "1.0.0"
  contact:
    name: RideShare Team
  license:
    name: MIT

servers:
  - url: http://localhost:8080
    description: Local development

tags:
  - name: Auth
    description: Registration and login
  - name: Rides
    description: Ride lifecycle
  - name: Users
    description: User profiles and driver availability
  - name: Locations
    description: Driver GPS and nearby search
  - name: Payments
    description: Ride payment processing
  - name: Notifications
    description: Real-time WebSocket events
  - name: Operations
    description: Health and metrics

# ── Security ──────────────────────────────────────────────────
security:
  - bearerAuth: []

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          example: "invalid ride ID"
        error_code:
          type: string
          example: "VALIDATION_ERROR"

    RegisterRequest:
      type: object
      required: [email, password, name, role]
      properties:
        email:
          type: string
          format: email
          example: "rider@example.com"
        password:
          type: string
          minLength: 8
          example: "securepass123"
        name:
          type: string
          maxLength: 100
          example: "Jane Doe"
        phone:
          type: string
          maxLength: 20
          example: "+1234567890"
        role:
          type: string
          enum: [rider, driver]
        vehicle_type:
          type: string
          description: Required if role=driver
          example: "sedan"
        license_plate:
          type: string
          description: Required if role=driver
          example: "ABC-1234"

    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string

    AuthResponse:
      type: object
      properties:
        token:
          type: string
          description: JWT access token
        user:
          $ref: "#/components/schemas/User"

    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
        name:
          type: string
        phone:
          type: string
        role:
          type: string
          enum: [rider, driver]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    UserWithDriver:
      allOf:
        - $ref: "#/components/schemas/User"
        - type: object
          properties:
            driver:
              $ref: "#/components/schemas/Driver"

    Driver:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        vehicle_type:
          type: string
        license_plate:
          type: string
        available:
          type: boolean
        rating:
          type: number
          format: float

    RideRequest:
      type: object
      required: [pickup_lat, pickup_long, dropoff_lat, dropoff_long]
      properties:
        pickup_lat:
          type: number
          minimum: -90
          maximum: 90
          example: 29.4241
        pickup_long:
          type: number
          minimum: -180
          maximum: 180
          example: -98.4936
        dropoff_lat:
          type: number
          minimum: -90
          maximum: 90
          example: 29.4500
        dropoff_long:
          type: number
          minimum: -180
          maximum: 180
          example: -98.5200

    Ride:
      type: object
      properties:
        id:
          type: string
          format: uuid
        rider_id:
          type: string
          format: uuid
        driver_id:
          type: string
          format: uuid
          nullable: true
        pickup_lat:
          type: number
        pickup_long:
          type: number
        dropoff_lat:
          type: number
        dropoff_long:
          type: number
        status:
          type: string
          enum: [requested, accepted, in_progress, completed, cancelled]
        fare:
          type: number
          description: Fare in USD
        surge_multiplier:
          type: number
        estimated_eta:
          type: integer
          description: ETA in seconds
        distance_km:
          type: number
        started_at:
          type: string
          format: date-time
          nullable: true
        ended_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    RideResponse:
      allOf:
        - $ref: "#/components/schemas/Ride"
        - type: object
          properties:
            matched_driver:
              $ref: "#/components/schemas/NearbyDriver"

    NearbyDriver:
      type: object
      properties:
        driver_id:
          type: string
          format: uuid
        distance_km:
          type: number
        lat:
          type: number
        long:
          type: number

    Payment:
      type: object
      properties:
        id:
          type: string
          format: uuid
        ride_id:
          type: string
          format: uuid
        amount:
          type: number
        currency:
          type: string
          example: "USD"
        status:
          type: string
          enum: [pending, completed, failed, refunded]
        stripe_txn_id:
          type: string
        created_at:
          type: string
          format: date-time

    ChargeRequest:
      type: object
      required: [ride_id, amount]
      properties:
        ride_id:
          type: string
          format: uuid
        amount:
          type: number
          minimum: 0.01
          maximum: 10000

    UpdateProfileRequest:
      type: object
      properties:
        name:
          type: string
          maxLength: 100
        phone:
          type: string
          maxLength: 20
        vehicle_type:
          type: string
          maxLength: 50
          description: Driver only
        license_plate:
          type: string
          maxLength: 20
          description: Driver only

    AvailabilityRequest:
      type: object
      required: [available]
      properties:
        available:
          type: boolean

    LocationUpdate:
      type: object
      required: [lat, long]
      properties:
        lat:
          type: number
          minimum: -90
          maximum: 90
        long:
          type: number
          minimum: -180
          maximum: 180

    PaginatedResponse:
      type: object
      properties:
        data:
          type: array
          items: {}
        page:
          type: integer
        limit:
          type: integer
        total_count:
          type: integer
        total_pages:
          type: integer
        has_more:
          type: boolean

# ── Paths ─────────────────────────────────────────────────────
paths:
  # ── Operations ──
  /health:
    get:
      tags: [Operations]
      summary: Health check
      security: []
      responses:
        "200":
          description: All services healthy
        "503":
          description: One or more services degraded

  /metrics:
    get:
      tags: [Operations]
      summary: Prometheus metrics
      security: []
      responses:
        "200":
          description: Prometheus text format

  # ── Auth ──
  /v1/auth/register:
    post:
      tags: [Auth]
      summary: Register a new account
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: Account created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Email already registered

  /v1/auth/login:
    post:
      tags: [Auth]
      summary: Login with email and password
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials

  # ── Rides ──
  /v1/rides/request:
    post:
      tags: [Rides]
      summary: Request a new ride
      description: Creates a ride, calculates fare/ETA/surge, and attempts automatic driver matching.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RideRequest"
      responses:
        "201":
          description: Ride created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RideResponse"
        "403":
          description: Only riders can request rides
        "409":
          description: Already have an active ride

  /v1/rides/history:
    get:
      tags: [Rides]
      summary: Get ride history (paginated)
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
      responses:
        "200":
          description: Paginated ride history
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedResponse"

  /v1/rides/{id}:
    get:
      tags: [Rides]
      summary: Get ride details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Ride details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ride"
        "404":
          description: Ride not found

  /v1/rides/{id}/accept:
    post:
      tags: [Rides]
      summary: Accept a ride (driver)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Ride accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ride"

  /v1/rides/{id}/complete:
    post:
      tags: [Rides]
      summary: Complete a ride (driver)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Ride completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ride"

  /v1/rides/{id}/cancel:
    post:
      tags: [Rides]
      summary: Cancel a ride
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Ride cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ride"

  # ── Users ──
  /v1/users/{id}:
    get:
      tags: [Users]
      summary: Get user profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserWithDriver"
        "404":
          description: User not found
    put:
      tags: [Users]
      summary: Update own profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileRequest"
      responses:
        "200":
          description: Profile updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserWithDriver"
        "403":
          description: Can only update your own profile

  /v1/users/{id}/availability:
    put:
      tags: [Users]
      summary: Set driver availability
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AvailabilityRequest"
      responses:
        "200":
          description: Availability updated
        "400":
          description: Only drivers can set availability

  # ── Locations ──
  /v1/locations/update:
    post:
      tags: [Locations]
      summary: Update driver GPS position
      description: Only drivers can update their location.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LocationUpdate"
      responses:
        "200":
          description: Location updated
        "403":
          description: Only drivers can update location

  /v1/locations/nearby:
    get:
      tags: [Locations]
      summary: Find nearby drivers
      parameters:
        - name: lat
          in: query
          required: true
          schema:
            type: number
        - name: long
          in: query
          required: true
          schema:
            type: number
        - name: radius
          in: query
          schema:
            type: number
            default: 5
            description: Search radius in km (max 50)
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Paginated nearby drivers
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedResponse"

  # ── Payments ──
  /v1/payments/charge:
    post:
      tags: [Payments]
      summary: Charge for a completed ride
      description: Idempotent — charging an already-charged ride returns the existing payment.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChargeRequest"
      responses:
        "200":
          description: Payment processed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"

  /v1/payments/ride/{ride_id}:
    get:
      tags: [Payments]
      summary: Get payment for a ride
      parameters:
        - name: ride_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Payment details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
        "404":
          description: Payment not found

  # ── Notifications ──
  /v1/ws:
    get:
      tags: [Notifications]
      summary: WebSocket connection for real-time events
      description: |
        Upgrades to WebSocket. Authenticate via `token` query parameter.
        Receives ride status updates and notifications in real-time.
      security: []
      parameters:
        - name: token
          in: query
          required: true
          schema:
            type: string
          description: JWT token for authentication
      responses:
        "101":
          description: Switching to WebSocket protocol
        "401":
          description: Invalid or missing token